Seed script working as expected with fewer files
